Real merge commit (not squash) to heal the ancestry divergence the prior
squash promotion (#882) introduced. Resolves 40 conflicts by taking develop's
side (develop is the newer source of truth; qa's copies were stale from the
last promotion): CI workflows, Cargo.lock, decomposition crates
(tenant/observability-engine/relational-executor/planner/storage-ports),
postgres relational_pipeline, query compute_scheduler/native_ops/route_cost,
raptor common, traits/mod, and test/docs files. Honors develop's deletion of
dead ViperPipeline (src/storage/engines/viper/pipeline*.rs, PR #947).

Versions already aligned (0.2.2) on both branches; no qa-only content lost
(the #707 OpenSSL vendoring is absent from both tips — superseded).

+++ crates/horizontal/proximadb-recall-bench/tests/fp16_e2e.rs
@@ -234,7 +234,7 @@ fn fp16_recall_at_10_meets_lld_q13_cosine_gate() {
 
     let q13_gate = RecallSlo::lld_defaults().cosine.at_10;
     assert!(
-        row.mean_recall >= q13_gate as f32,
+        row.mean_recall >= q13_gate,
         "fp16 mean recall@{TOP_K} = {} fell below LLD §Q13 cosine gate {} \
          (this is the LLD's go/no-go threshold for shipping fp16)",
         row.mean_recall,
